entrysafe-backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# MongoDB
motor==3.6.0
pymongo==4.9.2

# Firebase Admin (for JWT verification)
firebase-admin==6.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# CORS
python-dotenv==1.0.1

# Date/Time
python-dateutil==2.9.0

# HTTP Client
httpx==0.27.2

# OpenAI
openai>=1.0.0
